requirements.txt
anyio==4.6.2.post1
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
outcome==1.3.0.post0
pycparser==2.22
PySocks==1.7.1
python-telegram-bot==21.7
pytube==15.0.0
selenium==4.21.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
wsproto==1.2.0
yt-dlp==2024.11.18
